+++ .gitignore
@@ -1,1 +1,2 @@
 /data/
+/node_modules
