+++ src/utilities/chart-report-calculation.js
@@ -1,21 +1,22 @@
 export function maxTemperatures(weatherData) {
     let maxTemperatures = [];
     
-    for (let i = 0; i < weatherData.length; i++) {
-        weatherData[i].forEach((value, key) => {
+    for (let data of weatherData) {
+        data.forEach((value, key) => {
             if (key === "Max TemperatureC") {
                 maxTemperatures.push(value);
             }
         });
     }
+    return maxTemperatures
 }
 
 export function minTemperatures(weatherData) {
    
     let minTemperatures = [];
    
-    for (let i = 0; i < weatherData.length; i++) {
-        weatherData[i].forEach((value, key) => {
+    for (let data of weatherData) {
+        data.forEach((value, key) => {
             if (key === "Min TemperatureC") {
                 minTemperatures.push(value);
             }
