+++ weatherman.js
@@ -1,4 +1,3 @@
 import { main } from "./src/main.js";
-
-await main(process.argv);
+await main();
